move filtering into its own file
